Regroupe deux versions de fichier en un fichier de caractéristiques
